feat(brokerage): fixed labels and moved code to proper place
