+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/index.d.ts
@@ -345,6 +345,7 @@ type MessagesType = {
   'copy.crypto_selection_empty.needtoown': "You'll need to own crypto first..."
   'copy.crypto_selection_empty.receive': 'Are you holding crypto on a different wallet? Does a friend want to send you some Bitcoin? Copy and paste or share your unique wallet addresses QR codes.'
   'copy.crypto_selection_empty.receivetitle': 'Receive from Another Wallet'
+  'copy.currency': 'Currency'
   'copy.date': 'Date'
   'copy.error.locked_withdraw_error': 'Your crypto will be available to be withdrawn within {days} days.'
   'copy.failed': 'Failed'
@@ -581,9 +582,7 @@ type MessagesType = {
   'lockbox.service.messages.unknown': 'An unknown error has occurred.'
   'lockboxrequired.tooltip': 'Hardware device is required'
   'modals.brokerage.bank_preview.disclaimer': 'This account can be used for buys, sells & withdrawals.'
-  'modals.brokerage.remove_bank.description': "You'e about to remove your {bankAccount}"
-  'modals.brokerage.remove_bank.back': 'Back'
-  'modals.brokerage.curency': 'Curency'
+  'modals.brokerage.remove_bank.description': "You're about to remove your {bankAccount}"
   'modal.createaccountlearn.info1': 'Stellar requires that all Stellar accounts hold a minimum balance of lumens, or XLM. This means you cannot send a balance out of your Stellar Wallet that would leave your Stellar Wallet with less than the minimum balance. This also means that in order to send XLM to a new Stellar account, you must send enough XLM to meet the minimum balance requirement.'
   'modal.createaccountlearn.info2': 'The current minimum balance requirement is 1 XLM.'
   'modal.createaccountlearn.info3': "You can read more information about Stellar's minimum balance requirement at"
