+++ packages/blockchain-wallet-v4-frontend/src/data/components/brokerage/model.ts
@@ -1,0 +1,6 @@
+import { WalletFiatType } from 'blockchain-wallet-v4/src/types'
+
+export const DEFAULT_METHODS = {
+  currency: 'EUR' as WalletFiatType,
+  methods: []
+}
